Include GDB in the build
Build GDB on native, Linux (cross), and Win32. It is not built for
Arm targets.

On Linux you must install the termcap libraries to build GDB. On
Ubuntu, it's the libncurses-dev package or some variant of it. Note
that you need the 32-bit version.

On recent versions of MacOS, the linker requires that
-search_paths_first be specified so that it properly finds static
libraries. An LDFLAGS variable was added for use with native
tools. It's possible it could make builds on older versions of OS X
fail.